LayoutTests/platform/mac/accessibility/aria-option-role-expected.txt

option 1
option 2
This tests that the aria 'option' role works as expected. That is, it becomes a static text element with no children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstChild.role is 'AXRole: AXStaticText'
PASS firstChild.title is 'AXTitle: option 1'
PASS secondChild.role is 'AXRole: AXStaticText'
PASS secondChild.description is 'AXDescription: label 2'
PASS firstChild.childrenCount is 0
PASS secondChild.childrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
